+++ src/components/SearchBar/SearchBar.styled.js
@@ -1,0 +1,4 @@
+import styled from 'styled-components';
+export const SearchTypeSelectorBox = styled.div`
+  margin-bottom: 30px;
+`;
