+++ src/helpers/createSearchUrl.js
@@ -1,8 +1,8 @@
-export function createSearchUrl(query) {
-  const searchWord = query.toLowerCase();
+export function createSearchUrl(queryWord, type = 'query') {
+  const searchWord = queryWord.toLowerCase();
   const encodedSearchTerm = encodeURIComponent(searchWord);
 
-  const searchUrl = `/search?query=${encodedSearchTerm}`;
+  const searchUrl = `/search?query=${encodedSearchTerm}&type=${type}`;
   //   navigate(`/search?type=${searchType}&term=${encodedSearchTerm}`);
   console.log('searchUrl', searchUrl);
   return searchUrl;
