+++ src/pages/SearchPage/SearchPage.jsx
@@ -1,27 +1,23 @@
-import SearchForm from 'components/SearchForm/SearchForm';
 import React, { useState } from 'react';
 import { recipesG } from 'gannaFakeData';
 import DishCard from 'components/DishCard/DishCard';
 import { useLocation } from 'react-router-dom';
-import { ContainerWrapper, Container, SearchBar } from './Searchpage.styled';
+import { ContainerWrapper, Container } from './Searchpage.styled';
 import { useSearchParams } from 'react-router-dom';
-import SearchTypeSelector from 'components/SearchTypeSelector/SearchTypeSelector';
+import SearchBar from 'components/SearchBar/SearchBar';
+import { createSearchUrl } from 'helpers/createSearchUrl';
 
 export default function SearchPage() {
   const location = useLocation();
   // const [searchParams, setSearchParams] = useSearchParams();
   // const searchQuery = searchParams.get('query');
-  const [searchParams] = useSearchParams();
+  const [searchParams, setSearchParams] = useSearchParams();
   const searchQueryF = searchParams.get('query');
-  console.log('in SP', searchQueryF);
+  console.log('in SP query', searchQueryF);
   //---------------------------
-  const [searchType, setSearchType] = useState('query');
+
   // const [searchQuery, setSearchQuery] = useState('');
   // const [searchIngredient, setSearchIngredient] = useState('');
-
-  const handleSearchTypeChange = value => {
-    setSearchType(value);
-  };
 
   // const handleQueryChange = event => {
   //   setSearchQuery(event.target.value);
@@ -30,30 +26,18 @@ export default function SearchPage() {
   // const handleIngredientChange = event => {
   //   setSearchIngredient(event.target.value);
   // };
+  const handleSubmit = query => {
+    console.log('Query in SearchBar', query);
 
-  const handleSubmit = event => {
-    event.preventDefault();
-    // отправка запроса на бекенд с использованием searchType, searchQuery или searchIngredient
+    // const nextQuery = query !== '' ? { query } : {};
+    // setSearchParams(nextQuery);
+    const searchUrl = createSearchUrl(query);
+    console.log('SearchUrl in SearchBar', searchUrl);
   };
   //----------------------------
   return (
     <ContainerWrapper>
-      <SearchForm green marginBottom />
-      <SearchBar onSubmit={handleSubmit}>
-        <SearchTypeSelector
-          selectedValue={searchType}
-          onChange={handleSearchTypeChange}
-        />
-        {/* {searchType === 'query' ? (
-          <input type="text" value={searchQuery} onChange={handleQueryChange} />
-        ) : (
-          <input
-            type="text"
-            value={searchIngredient}
-            onChange={handleIngredientChange}
-          />
-        )} */}
-      </SearchBar>
+      <SearchBar onSubm={handleSubmit} />
       <Container>
         {recipesG.map(recipe => (
           <DishCard key={recipe._id} location={location} recipe={recipe} />
